Fix grades and average values (int -> double)

+++ Students/Func.cpp
@@ -1,8 +1,8 @@
 #include "Func.h"
 
-int Average (int* grade)
+double Average (double* grade)
 {
-    int sum=0;
+    double sum=0;
     for (int i=0; i<5;i++)
     {
         sum=sum+grade[i];
@@ -10,18 +10,18 @@ int Average (int* grade)
     return sum/5;
 }
 
-int minGrade (vector<TStudent>& students)
+double minGrade (vector<TStudent>& students)
 {
-    int AvGrades[students.size()];
+    double AvGrades[students.size()];
     for (int i=0;i<students.size();i++)
     {
         AvGrades[i]=students[i].averageG;
     }
     sortM(AvGrades,students.size());
-    return AvGrades[int(floor(students.size()*0.4))];
+    return AvGrades[floor(students.size()*0.4)];
 }
 
-int* sortM (int* AvGrade, int n)
+double* sortM (double* AvGrade, int n)
 {
     for (int i=0;i<n;i++)
     {
@@ -29,7 +29,7 @@ int* sortM (int* AvGrade, int n)
         {
             if (AvGrade[i]<AvGrade[j])
             {
-                int buf = AvGrade[i];
+                double buf = AvGrade[i];
                 AvGrade[i] = AvGrade[j];
                 AvGrade[j] = buf;
             }
